# code/storage/api-docs/api-docs-central.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ComandaFlow Central API Documentation",
        "description": "Central domain API for managing tenants and system-wide operations",
        "contact": {
            "name": "ComandaFlow Support",
            "email": "support@comandaflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Central API - Automatically uses current URL"
        },
        {
            "url": "/swagger/central"
        }
    ],
    "paths": {
...
        "/api/v1/central/tenants/{id}": {
            "get": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Get tenant details",
                "description": "Returns detailed information about a specific tenant",
                "operationId": "getTenant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acme-corp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "acme-corp"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "domains": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "domain": {
                                                        "type": "string"
                                                    },
                                                    "tenant_id": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "database": {
                                            "type": "string",
                                            "example": "tenantacme-corp"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tenant not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Update tenant information",
                "description": "Updates tenant data (name, plan, settings, etc.). Does not update domains.",
                "operationId": "updateTenant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acme-corp"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "description": "Tenant data to update",
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "ACME Corporation Updated"
                                            },
                                            "plan": {
                                                "type": "string",
                                                "example": "enterprise"
                                            },
                                            "settings": {
                                                "properties": {
                                                    "timezone": {
                                                        "type": "string",
                                                        "example": "America/New_York"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tenant updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tenant updated successfully"
                                        },
                                        "tenant": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Delete a tenant",
                "description": "Deletes a tenant, its domains, and its database. THIS ACTION IS IRREVERSIBLE!",
                "operationId": "deleteTenant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acme-corp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tenant deleted successfully"
                                        },
                                        "tenant_id": {
                                            "type": "string",
                                            "example": "acme-corp"
                                        },
                                        "deleted_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found"
                    },
                    "500": {
                        "description": "Error deleting tenant"
                    }
                }
            }
        },
        "/api/v1/central/tenants/{id}/domains": {
            "post": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Add domain to tenant",
                "description": "Adds a new domain to an existing tenant",
                "operationId": "addTenantDomain",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "acme-corp"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "domain"
                                ],
                                "properties": {
                                    "domain": {
                                        "type": "string",
                                        "example": "newdomain.comandaflow.local"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Domain added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Domain added successfully"
                                        },
                                        "domain": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found"
                    },
                    "422": {
                        "description": "Validation error (domain already exists)"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}